+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+- Adding recursive object condition checking
+
 ## [0.2.0] - 2019-09-11
 
 - Adding shallow object matching